cleanup: reorder setParam function, uint256 -> uint in Policies
move setParam to top, make them look uniform

+++ packages/network-contracts/contracts/OperatorTokenomics/SponsorshipPolicies/VoteKickPolicy.sol
@@ -43,7 +43,8 @@ contract VoteKickPolicy is IKickPolicy, Sponsorship {
     // can't be flagged again right after a no-kick result
     mapping (address => uint) public protectionEndTimestamp;
 
-    function setParam(uint256) external {
+    function setParam(uint) external {
+
     }
 
     function getFlagData(address operator) override external view returns (uint flagData) {
